requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
web3==6.11.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
cryptography>=42.0.0
eth-account==0.9.0
eth-utils==2.2.2
python-multipart==0.0.6
jinja2==3.1.2 